cleanup - pass 1
