+++ common/__init__.py
@@ -2,3 +2,4 @@ from activation import actf
 from convolution import GraphConvolution
 from mask import mask
 from posf import position_weight
+from bucket_iterator import BucketIterator
