+++ common/bucket_iterator.py
@@ -1,12 +1,11 @@
 # -*- coding: utf-8 -*-
-
 import math
 import random
 import torch
 import numpy
 
 class BucketIterator(object):
-    def __init__(self, data, batch_size, sort_key='sdat_graph', shuffle=True, sort=False):
+    def __init__(self, data, batch_size, sort_key='text_indices', shuffle=True, sort=True):
         self.shuffle = shuffle
         self.sort = sort
         self.sort_key = sort_key
@@ -25,11 +24,6 @@ class BucketIterator(object):
         return batches
 
     def pad_data(self, batch_data):
-        batch_bert_text_indices = []
-        batch_bert_seg_indices = []
-        batch_bert_raw_indices = []
-        batch_bert_aspect_indices = []
-
         batch_context = []
         batch_aspect = []
         batch_text_indices = []
@@ -41,64 +35,31 @@ class BucketIterator(object):
         batch_sentic_graph = []
         batch_sdat_graph = []
         max_len = max([len(t[self.sort_key]) for t in batch_data])
-        max_text_len = max([len(t['text_indices']) for t in batch_data])
-        max_text_len = 80
-        max_pad = 0
         for item in batch_data:
-            text_bert_indices, bert_segments_ids, text_raw_bert_indices, aspect_bert_indices,\
             context, aspect, text_indices, context_indices, aspect_indices, left_indices,\
             polarity, dependency_graph, sentic_graph, sdat_graph = \
-                item['text_bert_indices'], item['bert_segments_ids'], item['text_raw_bert_indices'], item['aspect_bert_indices'],\
                 item['context'], item['aspect'], \
                 item['text_indices'], item['context_indices'], item['aspect_indices'], item['left_indices'],\
                 item['polarity'], item['dependency_graph'], item['sentic_graph'], item['sdat_graph']
-
-            bert_text_padding = [0] * (max_len - len(text_bert_indices) + max_pad)
-            bert_seg_padding = [0] * (max_len - len(bert_segments_ids) + max_pad)
-            bert_raw_padding = [0] * (max_len - len(text_raw_bert_indices) + max_pad)
-            bert_aspect_padding = [0] * (max_len - len(aspect_bert_indices) + max_pad)
-
             text_padding = [0] * (max_len - len(text_indices))
             context_padding = [0] * (max_len - len(context_indices))
             aspect_padding = [0] * (max_len - len(aspect_indices))
             left_padding = [0] * (max_len - len(left_indices))
-
-            #print('max_len:', max_len)
-
-            #print('bert_text_padding:', bert_text_padding)
-
-            batch_bert_text_indices.append(text_bert_indices)
-            batch_bert_seg_indices.append(bert_segments_ids)
-            batch_bert_raw_indices.append(text_raw_bert_indices)
-            batch_bert_aspect_indices.append(aspect_bert_indices)
-
             batch_context.append(context)
             batch_aspect.append(aspect)
-            batch_text_indices.append(text_indices)
-            batch_context_indices.append(context_indices)
-            batch_aspect_indices.append(aspect_indices)
-            batch_left_indices.append(left_indices)
+            batch_text_indices.append(text_indices + text_padding)
+            batch_context_indices.append(context_indices + context_padding)
+            batch_aspect_indices.append(aspect_indices + aspect_padding)
+            batch_left_indices.append(left_indices + left_padding)
             batch_polarity.append(polarity)
-            #batch_dependency_graph.append(numpy.pad(dependency_graph, 
-                #((0,max_len-len(text_indices)+max_pad),(0,max_len-len(text_indices)+max_pad)), 'constant'))
-            #batch_sentic_graph.append(numpy.pad(sentic_graph, 
-                #((0,max_len-len(text_indices)+max_pad),(0,max_len-len(text_indices)+max_pad)), 'constant'))
-            #batch_sdat_graph.append(numpy.pad(sdat_graph, 
-                #((0,max_text_len-len(text_indices)),(0,max_text_len-len(text_indices))), 'constant'))
-            batch_sdat_graph.append(sdat_graph)
-
-        #for i in range(len(batch_bert_text_indices)):
-            #print('batch_bert_aspect_indices:', batch_bert_text_indices[i])
-            #print('-'*20)
-            #print('sentic_graph:', batch_sdat_graph[i])
-            #print('='*20)
+            batch_dependency_graph.append(numpy.pad(dependency_graph, 
+                ((0,max_len-len(text_indices)),(0,max_len-len(text_indices))), 'constant'))
+            batch_sentic_graph.append(numpy.pad(sentic_graph, 
+                ((0,max_len-len(text_indices)),(0,max_len-len(text_indices))), 'constant'))
+            batch_sdat_graph.append(numpy.pad(sdat_graph, 
+                ((0,max_len-len(text_indices)),(0,max_len-len(text_indices))), 'constant'))
 
         return { \
-                'text_bert_indices': torch.tensor(batch_bert_text_indices), \
-                'bert_segments_ids': torch.tensor(batch_bert_seg_indices), \
-                'text_raw_bert_indices': torch.tensor(batch_bert_raw_indices), \
-                'aspect_bert_indices': torch.tensor(batch_bert_aspect_indices), \
-
                 'context': batch_context, \
                 'aspect': batch_aspect, \
                 'text_indices': torch.tensor(batch_text_indices), \
