+++ infer_for_bert.py
@@ -8,7 +8,7 @@ import numpy as np
 
 from transformers import BertModel
 #from pytorch_pretrained_bert import BertModel
-from data_utils_bert import build_tokenizer, build_embedding_matrix, Tokenizer4Bert, ABSADataset, pad_and_truncate
+from data_utils_bert import Tokenizer4Bert, pad_and_truncate
 from models import SenticGCN_BERT
 from generate_sentic_dependency_graph import load_sentic_word, dependency_adj_matrix
